requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy[asyncio]==2.0.29
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
httpx==0.27.0
aiogram==3.4.1
passlib[bcrypt]==1.7.4
fastapi-admin==1.0.4
pytest==8.2.0
pytest-asyncio==0.23.6
black==24.4.2
isort==5.13.2
flake8==7.0.0
pre-commit==3.7.0
slowapi==0.1.9
redis==5.0.3
